Gradle: Ignore BOM dependencies imported from Maven
BOM dependencies do not define actual dependencies but version
constraints for other dependencies [1]. Therefore these dependencies can
be ignored.

Before this change BOM dependencies appeared in the dependency tree.
Additionally all version constraints defined by the BOM which were used
inside the same module also appeared as dependencies of the BOM. For
larger projects this led to problems with the optimizations introduced
in 1234567 because BOM dependencies could have a lot of different
subtrees which resulted in a lot of tree comparisons. Also, this led to
a huge increase in the size of the dependency tree, because artifacts
published with proper BOM support would in turns have a dependency to
the BOM [2], which then in turn would again reference the artifacts.

For one larger test project using multiple BOMs the above issued caused
the analyzer to run for 14 hours until the machine went out of memory.
After this change the analysis of the same project takes about one
minute.

[1] https://docs.gradle.org/current/userguide/platforms.html#sub:bom_import
[2] https://blog.gradle.org/alignment-with-gradle-module-metadata

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>